pyproject.toml
[project]
name = "bookery"
version = "0.1.0"
description = "Django CRUD application for managing books"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-extensions>=3.2.3",
    "django-filter>=24.3",
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
]
